vits_nepali/requirements.txt
# Placeholder file
# requirements.txt
torch>=1.13.0
torchaudio>=0.13.0
numpy>=1.21.0
librosa>=0.9.0
pyyaml>=6.0
tqdm>=4.64.0
matplotlib>=3.5.0